Fix WPF crash when switching a transaction to Remove or Withdraw
My working theory is that WPF got confused when property change events were raised in an order such that the transaction shouldn't belong to this account at all given intermediate state.
It's evidently important to set another property to bind this transaction to this account before clearing the other property that already bound the transaction to this account.